requirements.txt
torch
torchvision
torchaudio
pillow
requests